Initial draft of a graduated supply cap
The cap should follow a 100, 250, 500, 1000 BTC monthly cap schedule, in
addition to a 1-day 1 BTC grace period after the system launches before
more deposits allowed.

The cap is calculated by the TBTC token's total supply. If depositors
instead hold on to TDTs without minting TBTC, more BTC can be deposited
into the system before the cap is reached. An alternative approach would
be to keep track of active TDTs, but it's an expensive alternative with
a greater likelihood of abuse.

Still up: updated tests for getAllowNewDeposits().